services/recording/historical_requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
Deprecated==1.2.13
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.11.2
eth-abi==3.0.0
eth-hash==0.3.2
eth-typing==3.0.0
eth-utils==2.0.0
fastapi==0.78.0
frozenlist==1.3.0
h11==0.13.0
idna==3.3
kombu==5.2.4
motor==3.0.0
multidict==6.0.2
parsimonious==0.8.1
prompt-toolkit==3.0.29
pycryptodome==3.14.1
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toolz==0.11.2
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.7.2
